# openapi.yaml
openapi: 3.0.3
info:
  title: Grant Application Portal API
  version: 1.0.0
  description: >
    Backend for the grant application portal: applicants browse grants, fill in
    a form generated from per-grant requirements, upload documents and submit;
    staff define grants, review applications and export the results.
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        details: {}
    User:
      type: object
      properties:
        id: { type: string, description: 64-bit id serialized as a string }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [applicant, admin, reviewer] }
        organization: { type: string, nullable: true }
...
    ApplicationStatus:
      type: string
      enum: [draft, submitted, under_review, approved, rejected, withdrawn]
    ValidationIssue:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        message: { type: string }
    Progress:
      type: object
      properties:
        total: { type: integer }
        completed: { type: integer }
        requiredTotal: { type: integer }
        requiredCompleted: { type: integer }
        percent: { type: integer }
        missing:
          type: array
          items: { type: string }
    ApplicationSummary:
      type: object
      properties:
        id: { type: string }
        status: { $ref: '#/components/schemas/ApplicationStatus' }
        status_label: { type: string }
        submitted_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time }
        grant:
          type: object
          properties:
            id: { type: string }
            title: { type: string }
            deadline: { type: string, format: date-time, nullable: true }
        applicant: { $ref: '#/components/schemas/User' }
        review_count: { type: integer }
    ApplicationDetail:
      allOf:
        - $ref: '#/components/schemas/ApplicationSummary'
        - type: object
          properties:
            decision_note: { type: string, nullable: true }
            values:
              type: object
              additionalProperties: {}
              description: One entry per requirement key.
            files:
              type: object
              additionalProperties:
                type: object
                nullable: true
                properties:
                  id: { type: string }
                  filename: { type: string }
                  size_bytes: { type: string }
                  mimetype: { type: string }
            progress: { $ref: '#/components/schemas/Progress' }
            average_score: { type: number, nullable: true }
            allowed_transitions:
              type: array
              items: { $ref: '#/components/schemas/ApplicationStatus' }
            validation_issues:
              type: array
              items: { $ref: '#/components/schemas/ValidationIssue' }
            reviews:
              type: array
              items: { $ref: '#/components/schemas/Review' }
    ResponseInput:
      type: object
      description: Identify the requirement by `key` or `requirement_id`.
      properties:
        requirement_id: { type: string }
        key: { type: string }
        value:
          description: Scalar, boolean or string array, matching the field type.
        file_id: { type: string, nullable: true }
    Review:
      type: object
      properties:
        id: { type: string }
        applicationId: { type: string }
        reviewerId: { type: string }
        score: { type: number, nullable: true }
        notes: { type: string, nullable: true }
        recommendation:
          type: string
          nullable: true
          enum: [approve, reject, needs_more_info]
        reviewer: { $ref: '#/components/schemas/User' }
...
  responses:
...
    Forbidden:
      description: The role is not allowed to perform this action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /applications:
    get:
      tags: [Applications]
      summary: Your applications, or every application for staff
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: grant_id, schema: { type: string } }
        - { in: query, name: q, schema: { type: string }, description: Staff only }
        - in: query
          name: sort
          schema: { type: string, enum: [created_at, updated_at, submitted_at, status], default: updated_at }
        - { in: query, name: order, schema: { type: string, enum: [asc, desc], default: desc } }
        - { in: query, name: take, schema: { type: integer, default: 25 } }
        - { in: query, name: skip, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ApplicationSummary' }
                  total: { type: integer }
    post:
      tags: [Applications]
      summary: Start a draft application
      description: Returns the existing draft when one already exists for this grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grant_id]
              properties:
                grant_id: { type: string }
      responses:
        '200': { description: An application already existed }
        '201': { description: Draft created }
        '400': { description: The grant is closed, archived or not open yet }
  /applications/export.csv:
    get:
      tags: [Applications]
      summary: CSV export of the filtered list (staff)
      description: >
        Scoping the export to one grant adds a column per requirement key.
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: grant_id, schema: { type: string } }
        - { in: query, name: q, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            text/csv:
              schema: { type: string }
        '403': { $ref: '#/components/responses/Forbidden' }
  /applications/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Applications]
      summary: One application with answers, progress and reviews
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
    patch:
      tags: [Applications]
      summary: Change the status (submit, withdraw, review, decide)
      description: >
        Allowed transitions depend on the role. Submitting re-validates every
        answer server-side and rejects an incomplete application with the list
        of issues in `details`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/ApplicationStatus' }
                note: { type: string, description: Included in the applicant's notification }
      responses:
        '200': { description: Updated }
        '400':
          description: The application is not complete
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409': { description: The transition is not allowed, or the deadline has passed }
  /applications/{id}/responses:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    put:
      tags: [Applications]
      summary: Upsert answers (used by autosave)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/ResponseInput' }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  saved: { type: integer }
                  progress: { $ref: '#/components/schemas/Progress' }
                  validation_issues:
                    type: array
                    items: { $ref: '#/components/schemas/ValidationIssue' }
        '409': { description: Already submitted, or the deadline has passed }
  /applications/{id}/pdf:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Applications]
      summary: Download the application as a PDF
      responses:
        '200':
          description: OK
          content:
            application/pdf:
              schema: { type: string, format: binary }
  /applications/{id}/remind:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      tags: [Applications]
      summary: Notify the applicant about missing items (staff)
      responses:
        '200': { description: Sent }
...